Fix issues and debug comments

+++ app/src/main/java/com/example/weatherapp/weatherResponseData/ForecastWeatherResponseApi.kt
@@ -51,7 +51,7 @@ data class ForecastWeatherResponseApi(
         @SerializedName("main")
         val main: Main?,
         @SerializedName("pop")
-        val pop: Int?,
+        val pop: Double?,
         @SerializedName("sys")
         val sys: Sys?,
         @SerializedName("visibility")
